+++ src/main/java/net/silthus/skills/events/SetPlayerLevelEvent.java
@@ -16,11 +16,11 @@ public class SetPlayerLevelEvent extends PlayerEvent implements Cancellable {
     private static final HandlerList handlerList = new HandlerList();
 
     private final PlayerLevel level;
-    private final long oldLevel;
-    private long newLevel;
+    private final int oldLevel;
+    private int newLevel;
     private boolean cancelled;
 
-    public SetPlayerLevelEvent(PlayerLevel level, long oldLevel, long newLevel) {
+    public SetPlayerLevelEvent(PlayerLevel level, int oldLevel, int newLevel) {
         super(level.player());
         this.level = level;
         this.oldLevel = oldLevel;
